backend/requirements.txt
fastapi
uvicorn
openai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
dotenv
asyncpg
sqlalchemy
aembic
psycopg2-binary